+++ frameworks/solid/src/components/Form/Form.tsx
@@ -2,8 +2,8 @@ import {
   INTERNAL,
   type Schema,
   type SubmitHandler,
-  validateFormInput,
 } from '@formisch/core/solid';
+import { handleSubmit } from '@formisch/methods/solid';
 import { type JSX, splitProps } from 'solid-js';
 import type { FormStore } from '../../types/index.ts';
 
@@ -30,39 +30,7 @@ export function Form(props: FormProps): JSX.Element {
       ref={(element) => {
         props.of[INTERNAL].element = element;
       }}
-      onSubmit={async (event) => {
-        // Prevent default behavior of browser
-        event.preventDefault();
-
-        // Get internal form store
-        const internalFormStore = props.of[INTERNAL];
-
-        // Update submit state of form
-        internalFormStore.isSubmitted.value = true;
-        internalFormStore.isSubmitting.value = true;
-
-        // Try to run submit actions if form is valid
-        try {
-          const result = await validateFormInput(internalFormStore, {
-            shouldFocus: true,
-          });
-          if (result.success) {
-            await props.onSubmit(result.output, event);
-          }
-
-          // If an error occurred, set form errors
-        } catch (error) {
-          internalFormStore.errors.value = [
-            error instanceof Error
-              ? error.message
-              : 'An unknown error has occurred.',
-          ];
-
-          // Finally set submitting back to "false"
-        } finally {
-          internalFormStore.isSubmitting.value = false;
-        }
-      }}
+      onSubmit={(event) => handleSubmit(props.of, props.onSubmit)(event)}
     />
   );
 }