+++ frameworks/vue/src/components/Form/Form.vue
@@ -1,10 +1,6 @@
 <script setup lang="ts" generic="TSchema extends Schema = Schema">
-import {
-  INTERNAL,
-  Schema,
-  SubmitHandler,
-  validateFormInput,
-} from '@formisch/core/vue';
+import { Schema, SubmitHandler } from '@formisch/core/vue';
+import { handleSubmit } from '@formisch/methods/vue';
 import { FormStore } from '../../types';
 
 export type FormProps<TSchema extends Schema = Schema> = {
@@ -14,39 +10,13 @@ export type FormProps<TSchema extends Schema = Schema> = {
 
 const props = defineProps<FormProps<TSchema>>();
 
-async function handleSubmit(event: Event) {
-  event.preventDefault();
-  // Get internal form store
-  const internalFormStore = props.of[INTERNAL];
-
-  // Update submit state of form
-  internalFormStore.isSubmitted.value = true;
-  internalFormStore.isSubmitting.value = true;
-
-  // Try to run submit actions if form is valid
-  try {
-    const result = await validateFormInput(internalFormStore, {
-      shouldFocus: true,
-    });
-    if (result.success) {
-      await props.onSubmit(result.output, event as SubmitEvent);
-    }
-
-    // If an error occurred, set form errors
-  } catch (error) {
-    internalFormStore.errors.value = [
-      error instanceof Error ? error.message : 'An unknown error has occurred.',
-    ];
-
-    // Finally set submitting back to "false"
-  } finally {
-    internalFormStore.isSubmitting.value = false;
-  }
-}
+const handler = handleSubmit(props.of, props.onSubmit) as (
+  event: Event
+) => void;
 </script>
 
 <template>
-  <form novalidate @submit="handleSubmit">
+  <form novalidate @submit="handler">
     <slot></slot>
   </form>
 </template>