+++ frameworks/qwik/CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Add `SubmitHandler` type from `@formisch/core` to exports
-- Refactor useField and useFieldArray to improve implementation
+- Refactor `useField` and `useFieldArray` to improve implementation
+- Refactor `Form` component to use new `handleSubmit` method
 - Fix bug when resetting dynamic field array items
 
 ## v0.1.1 (July 17, 2025)